Change locations to avoid cyclic import
