+++ sktime/benchmarking/__init__.py
@@ -1,5 +1,5 @@
 """Benchmarking module."""
 
-from sktime.benchmarking._storage_handlers import load_results_to_dataframe
+from sktime.benchmarking._results_persistence import load_results_to_dataframe
 
 __all__ = ["load_results_to_dataframe"]
