+++ sktime/benchmarking/_storage_handlers.py
@@ -495,36 +495,6 @@ def results_to_dataframe(results: list[ResultObject]) -> pd.DataFrame:
     return df
 
 
-def load_results_to_dataframe(path: str | Path) -> pd.DataFrame:
-    """Load persisted benchmark results into a summary pandas DataFrame.
-
-    Reads a completed results file (``.json``, ``.csv``, or ``.parquet``).
-    When the final file is absent, partial checkpoints from an interrupted
-    run are loaded from ``{path}.parts/`` if present.
-
-    Parameters
-    ----------
-    path : str or pathlib.Path
-        Path to the benchmark results file. Must refer to a file, not a
-        directory. Supported extensions are ``.json``, ``.csv``, and
-        ``.parquet``.
-
-    Returns
-    -------
-    pandas.DataFrame
-        Aggregated benchmark metrics per task-model pair. Empty when no
-        results are found at ``path`` or in its checkpoint directory.
-
-    Raises
-    ------
-    ValueError
-        If no storage handler supports the file extension of ``path``.
-    """
-    from sktime.benchmarking._results_persistence import BenchmarkResultsPersistence
-
-    return results_to_dataframe(BenchmarkResultsPersistence(str(path)).load())
-
-
 def get_storage_backend(path: str | Path) -> BaseStorageHandler:
     """Return the storage handler for a results file path.
